# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0",
    "title": "presio"
  },
  "servers": [
    {
      "url": "https://localhost:5001/api/"
    },
    {
      "url": "http://localhost:8080/",
      "description": ""
    }
  ],
  "paths": {
    "/auth/signUp": {
      "post": {
        "summary": "Create New User",
        "description": "Create a new user.",
        "operationId": "post-user",
        "requestBody": {
          "$ref": "#/components/requestBodies/SignUp"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/User"
          },
          "400": {
            "description": "Missing Required Information"
          },
          "409": {
            "description": "Email Already Taken"
          }
        },
        "servers": [
          {
            "url": "https://localhost:5001/api/"
          },
          {
            "url": "http://localhost:8080/",
            "description": ""
          }
        ]
      }
    },
    "/auth/signIn": {
      "post": {
        "summary": "",
        "operationId": "post-signIn",
        "requestBody": {
          "$ref": "#/components/requestBodies/SignIn"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/User"
          }
        },
        "servers": [
          {
            "url": "https://localhost:5001/api/"
          },
          {
            "url": "http://localhost:8080/",
            "description": ""
          }
        ]
      }
    },
...
  },
  "components": {
...
    "requestBodies": {
      "SignUp": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        }
      },
      "SignIn": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        }
      },
...
    },
    "responses": {
      "User": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "email",
                "password"
              ]
            }
          }
        }
      },
...
    },
...
  }
}